Update find_scene.py to use orbits and scene modules and move to nci/preparation
